refactor: CompositionalLayout 타입으로 분리
